// apps/desktop/src/ipc/generated/FilesListing.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { FolderDto } from "./FolderDto";
import type { ProjectFileDto } from "./ProjectFileDto";

export type FilesListing = { folders: Array<FolderDto>, files: Array<ProjectFileDto>, };